backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]
python-jose
razorpay
httpx
websockets
python-multipart
